added path variable to resolve nested folders for sections or areas in the angular routes
The views need to have the same folder structure as each of the controllers, though.
The update is non breaking since the path has a default value set to empty string if nothing is passed through.
The path passed in has to have the trailing "/" (see the example route in the app.js file).